requirements.txt
Flask==3.0.3
gunicorn==23.0.0
redis==6.4.0
requests==2.32.5
rq==2.6.0
youtube-transcript-api==1.2.2
yt-dlp==2025.9.5
python-dotenv==1.1.1
beautifulsoup4
feedparser
faster-whisper
lxml